--- tcpip/project/src/main/java/kyungmin/tcpip/project/client/ASD.java
@@ -1,7 +1,0 @@
-package kyungmin.tcpip.project.client;
-
-public class ASD {
-
-
-
-}
